Address review round 2: cover the append proof; split kani module
- Add a `kani::cover!(valid_up_to > 0, ...)` to
  `append_valid_prefix_matches_checked_decode` so the meaningful case — a
  non-empty prefix that actually reaches the `unsafe from_utf8_unchecked`
  call — is proven reachable, rather than only the `valid_up_to == 0` early
  return (CodeRabbit). The cover is SATISFIED and the proof still verifies.
- Move the `#[cfg(kani)]` proof module out of `utf8.rs` into a `#[path]`
  sibling `utf8_kani_proofs.rs`. `utf8.rs` was at the 400-line module cap, so
  the added cover would have exceeded it; the split gives lasting headroom
  and keeps both files well under the limit.
